+++ cypress/e2e/registration_files.spec.ts
@@ -16,6 +16,6 @@ describe('Registration: File upload', () => {
 
     cy.fixture('img.jpg').as('file');
     cy.get('input[type=file]').first().selectFile('@file', { force: true });
-    cy.get('[data-testid=uploaded-file-row]').should('be.visible');
+    cy.get('[data-testid=uploaded-file-card]').should('be.visible');
   });
 });
