+++ package.json
@@ -37,7 +37,7 @@
     "@mui/icons-material": "^5.10.15",
     "@mui/lab": "5.0.0-alpha.109",
     "@mui/material": "^5.10.15",
-    "@mui/x-date-pickers": "^5.0.9",
+    "@mui/x-date-pickers": "^5.0.8",
     "@reduxjs/toolkit": "^1.9.0",
     "@uppy/aws-s3-multipart": "^1.8.18",
     "@uppy/core": "^1.20.1",
