+++ src/components/TabButton.tsx
@@ -13,7 +13,6 @@ export const TabButton = ({ isSelected, children, ...props }: TabButtonProps) =>
       fontSize: '1.2rem',
       color: isSelected ? 'primary' : 'text.primary',
       borderBottom: isSelected ? '0.2rem solid' : null,
-      textTransform: 'none',
     }}
     {...props}>
     {children}
