+++ src/pages/basic_data/app_admin/InstitutionList.tsx
@@ -1,16 +1,6 @@
 import { useTranslation } from 'react-i18next';
 import { Link as RouterLink } from 'react-router-dom';
-import {
-  Button,
-  Paper,
-  Table,
-  TableBody,
-  TableCell,
-  TableContainer,
-  TableHead,
-  TableRow,
-  Typography,
-} from '@mui/material';
+import { Button, Table, TableBody, TableCell, TableContainer, TableHead, TableRow, Typography } from '@mui/material';
 import EditIcon from '@mui/icons-material/Edit';
 import { visuallyHidden } from '@mui/utils';
 import { SimpleCustomerInstitution } from '../../../types/customerInstitution.types';
@@ -26,14 +16,22 @@ export const InstitutionList = ({ institutions }: InstitutionListProps) => {
   const { t } = useTranslation();
 
   return (
-    <TableContainer component={Paper}>
+    <TableContainer>
       <Table data-testid={dataTestId.basicData.customers.customerList} sx={alternatingTableRowColor}>
-        <caption style={visuallyHidden}>{t('basic_data.institutions.admin_institutions')}</caption>
+        <caption>
+          <span style={visuallyHidden}>{t('basic_data.institutions.admin_institutions')}</span>
+        </caption>
         <TableHead>
           <TableRow>
-            <TableCell>{t('common.name')}</TableCell>
-            <TableCell>{t('common.date')}</TableCell>
-            <TableCell>{t('common.actions')}</TableCell>
+            <TableCell>
+              <Typography fontWeight="bold">{t('common.name')}</Typography>
+            </TableCell>
+            <TableCell>
+              <Typography fontWeight="bold">{t('common.date')}</Typography>
+            </TableCell>
+            <TableCell>
+              <Typography fontWeight="bold">{t('common.actions')}</Typography>
+            </TableCell>
           </TableRow>
         </TableHead>
         <TableBody>
