+++ src/pages/basic_data/institution_admin/person_register/PersonRegisterPage.tsx
@@ -1,5 +1,4 @@
 import {
-  Paper,
   Skeleton,
   Table,
   TableBody,
@@ -81,15 +80,25 @@ export const PersonRegisterPage = () => {
         <Typography>{t('basic_data.person_register.no_employees_found')}</Typography>
       ) : (
         <>
-          <TableContainer component={Paper}>
+          <TableContainer>
             <Table size="small" sx={alternatingTableRowColor}>
               <caption style={visuallyHidden}>{t('basic_data.person_register.employee_table_caption')}</caption>
               <TableHead>
                 <TableRow>
-                  <TableCell>{t('basic_data.person_register.person_id')}</TableCell>
-                  <TableCell>{t('basic_data.person_register.national_identity_number')}</TableCell>
-                  <TableCell>{t('common.name')}</TableCell>
-                  <TableCell>{t('common.employments')}</TableCell>
+                  <TableCell>
+                    <Typography fontWeight="bold">{t('basic_data.person_register.person_id')}</Typography>
+                  </TableCell>
+                  <TableCell>
+                    <Typography fontWeight="bold">
+                      {t('basic_data.person_register.national_identity_number')}
+                    </Typography>
+                  </TableCell>
+                  <TableCell>
+                    <Typography fontWeight="bold">{t('common.name')}</Typography>
+                  </TableCell>
+                  <TableCell>
+                    <Typography fontWeight="bold">{t('common.employments')}</Typography>
+                  </TableCell>
                   <TableCell />
                 </TableRow>
               </TableHead>
