+++ src/utils/testfiles/mockRegistration.ts
@@ -9,7 +9,7 @@ import { mockUser } from './mock_feide_user';
 
 export const mockRegistration: JournalRegistration = {
   type: 'Publication',
-  id: 'https://api.dev.nva.aws.unit.no/registration/12345679',
+  id: 'https://frontend.dev.nva.aws.unit.no/registration/12345679',
   identifier: '12345679',
   createdDate: new Date(2020, 1).toISOString(),
   modifiedDate: new Date(2020, 2).toISOString(),
